Move parse_response to temba.utils.http
