+++ temba/utils/http.py
@@ -1,3 +1,8 @@
+from http.client import HTTPResponse
+from io import BytesIO
+
+import urllib3
+
 from django.conf import settings
 
 
@@ -24,3 +29,21 @@ class HttpEvent(object):
 
     def __str__(self):  # pragma: no cover
         return "%s %s %s %s %s" % (self.method, self.url, self.status_code, self.response_body, self.request_body)
+
+
+def parse_response(data):
+    """
+    Parses a saved HTTP response trace, e.g. "HTTP/1.1 200 OK\r\n\r\n{\"errors\":[]}"
+    """
+
+    class BytesIOSocket:
+        def __init__(self, content):
+            self.handle = BytesIO(content)
+
+        def makefile(self, mode):
+            return self.handle
+
+    response = HTTPResponse(BytesIOSocket(data.encode("utf-8")))
+    response.begin()
+
+    return urllib3.HTTPResponse.from_httplib(response)
